moved backend initialization code to app.js
